Clips a point layer by a filtered polygon layer

Returns the clipped point layer containing user specified columns from point layer as well as column indicating associated value used to filter plygon layer.

Usage

fpr_dbq_clip(
  schtab_pnt = NULL,
  schtab_pol = NULL,
  join_column = NULL,
  schtab_pol_filter = NULL,
  cols_schtab_pnt = "*"
)

Arguments

schtab_pnt

String (quoted) name of point layer schema.table.

schtab_pol

String (quoted) name of polygon layer schema.table.

join_column

String (quoted) name of column to join tables on from polygon. See column names of any table with fpr_dbq_lscols.

schtab_pol_filter

String (quoted) or quoted comma seperated vector that defines how the polygon table is filtered, example c("FRAN", "NECR").

cols_schtab_pnt

String (quoted) or list of column names to select from point layer, default is all columns (*). Must include the geom column if user wants an sf object.

Value

Sf of clipped point layer

See also

Examples

if (FALSE) fpr_db_query(query = fpr_dbq_clip('whse_environmental_monitoring.envcan_hydrometric_stn_sp',
'whse_basemapping.fwa_watershed_groups_poly', 'watershed_group_code', c("FRAN", "NECR")))